Major rewrite of this project.
Updated to v2.0 of the component field.
Updated to v2.0 of the translations field.
Changed the data structure of the json data.
Updated the middleware to better handle locales.
Changed to that if there is only one locale, no locale is used in the
url.
Updated the examples for this project, to match the new structure.